Simplificación de funciones de Creación de Pokémon
- Mejora de aleatoriedad del juego.
- Sistema de huevos simplificado (notas de desarrollo, chequear).

+++ include/field_effect.h
@@ -46,7 +46,7 @@ void SpriteCB_AshLaunch(struct Sprite *);
 
 void MultiplyPaletteRGBComponents(u32 i, u8 r, u8 g, u8 b);
 void FreeResourcesAndDestroySprite(struct Sprite *sprite, u8 spriteId);
-u8 CreateMonSprite_PicBox(u16 species, s16 x, s16 y, u8 subpriority);
+u32 CreateMonSprite_PicBox(u32 species, s32 x, s32 y, u32 subpriority);
 void StartEscapeRopeFieldEffect(void);
 
 #endif // GUARD_FIELD_EFFECTS_H
